Divided the component into small files
