+++ src/Components/ReviewBooking/FareDetails.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const FareDetails = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default FareDetails
